requirements.txt
pandas==3.0.1
Jinja2==3.1.6
load-dotenv==0.1.0
openpyxl==3.1.5